+++ testdata/de/dm/intellij/liferay/language/osgi/ReferenceUnbindReferenceContributorTest/org/osgi/service/component/annotations/Reference.java
@@ -1,0 +1,7 @@
+package org.osgi.service.component.annotations;
+
+public @interface Reference {
+
+	String unbind() default "";
+
+}