+++ changelog.md
@@ -3,6 +3,9 @@ Liferay Plugin for IntelliJ
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.25
+- Feature: code completion for the `unbind` attribute in a `@Reference` annotation
+
 ## 0.0.24
 - Feature: add deprecation checks for Liferay 2024.Q4
 - Feature: add support for IntelliJ IDEA 2025.1, dropped support for older versions